pyproject.toml
[tool.poetry]
name = "leantree"
version = "1.0.0"
description = "REPL interaction and data extraction for Lean 4."
authors = ["Matěj Kripner <kripner@ufal.mff.cuni.cz>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.12"
numpy = ">=1.26.4"
tqdm = ">=4.67.0"
PrettyPrintTree = ">=2.0.1"
psutil = ">=6.1.0"

[tool.poetry.group.dev]
optional = true

[tool.poetry.group.dev.dependencies]
mypy = ">=1.11.2"
pytest = ">=8.3.3"
pytest-asyncio = ">=0.21.0"
requests = ">=2.32.3"
torch = ">=2.2.0"
scipy = ">=1.11.3"
transformers = ">=4.36.0"
datasets = ">=2.14.5"

[tool.poetry.build]
script = "scripts/build_repl.py"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
